# contracts/openapi/identity.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autumn Wind Ai Identity Service API",
    "version": "1.0.0",
    "description": "Identity Service 的注册、认证、会话、用户管理和认证策略契约。"
  },
  "servers": [
    {
      "url": "/",
      "description": "由 API Gateway 路由的相对地址"
    }
  ],
...
  "paths": {
    "/api/v1/auth/csrf": {
      "get": {
        "tags": ["公共认证"],
        "summary": "获取 CSRF 防护信息",
        "operationId": "getCsrfProtection",
        "security": [],
        "responses": {
          "200": {
            "description": "已建立 CSRF 防护上下文。",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "./common.openapi.json#/components/headers/CorrelationId"
              },
              "X-CSRF-TOKEN": {
                "description": "与 HttpOnly CSRF Cookie 对应的掩码 Token，用于后续同源状态变更请求。",
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                }
              },
              "Set-Cookie": {
                "description": "建立 host-only AW_CSRF Cookie，属性为 HttpOnly、Secure、SameSite=Lax 和 Path=/，不设置 Domain。",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfProtection"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/auth/registration-options": {
      "get": {
        "tags": ["公共认证"],
        "summary": "读取公开注册选项",
        "description": "只返回注册页面需要的公开策略，不暴露邮箱域名单和登录锁定参数。",
        "operationId": "getRegistrationOptions",
        "security": [],
        "responses": {
          "200": {
            "description": "当前公开注册选项。",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "./common.openapi.json#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationOptionsView"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/auth/registrations": {
      "post": {
        "tags": ["公共认证"],
        "summary": "注册账户",
        "operationId": "registerUser",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/Accepted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/v1/auth/sessions": {
      "post": {
        "tags": ["公共认证"],
        "summary": "登录并创建会话",
        "operationId": "createSession",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功；会话凭据通过受保护 Cookie 设置。",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "./common.openapi.json#/components/headers/CorrelationId"
              },
              "Set-Cookie": {
                "description": "包含 HttpOnly、Secure、SameSite=Lax 和 Path=/ 的 host-only 不透明会话 Cookie；不设置 Domain，不得记录或向其他域转发。",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": ["公共认证"],
        "summary": "读取当前会话",
        "operationId": "getCurrentSession",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "当前有效会话。",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "./common.openapi.json#/components/headers/CorrelationId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["公共认证"],
        "summary": "注销当前会话",
        "operationId": "deleteCurrentSession",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "204": {
            "description": "注销成功并清除浏览器会话 Cookie。",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "./common.openapi.json#/components/headers/CorrelationId"
              },
              "Set-Cookie": {
                "description": "使用 Path=/、Max-Age=0、HttpOnly、Secure 和 SameSite=Lax 清除 host-only AW_SESSION Cookie。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "AW_SESSION",
        "description": "浏览器不透明会话凭据。"
      },
      "ServiceJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "包含明确 Issuer、Audience 和服务身份的短时 Service JWT。"
      }
    },
    "parameters": {
      "CsrfToken": {
        "name": "X-CSRF-TOKEN",
        "in": "header",
        "required": true,
        "description": "与当前 CSRF 防护上下文匹配的值。",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
...
    },
    "responses": {
      "Accepted": {
        "description": "请求已受理；响应不披露账户是否存在。",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "./common.openapi.json#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AcceptedView"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "请求字段或业务参数无效。",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "./common.openapi.json#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./common.openapi.json#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "认证失败或会话已失效。",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "./common.openapi.json#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./common.openapi.json#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "当前身份或账户状态不允许执行该操作。",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "./common.openapi.json#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./common.openapi.json#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "请求频率超过限制。",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "./common.openapi.json#/components/headers/CorrelationId"
          },
          "Retry-After": {
            "description": "再次尝试前需要等待的秒数。",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./common.openapi.json#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "所需身份能力尚未启用，请稍后重试或联系管理员。",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "./common.openapi.json#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./common.openapi.json#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "服务内部错误；客户端只应依赖稳定错误码和关联 ID，不得依赖内部诊断信息。",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "./common.openapi.json#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./common.openapi.json#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "CsrfProtection": {
        "type": "object",
        "additionalProperties": true,
        "required": ["headerName", "parameterName", "value"],
        "properties": {
          "headerName": {
            "type": "string"
          },
          "parameterName": {
            "type": "string"
          },
          "value": {
            "type": "string",
            "description": "仅用于同源状态变更请求，不得记录。"
          }
        }
      },
      "RegistrationOptionsView": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "publicRegistrationEnabled",
          "emailVerificationRequired",
          "passwordMinimumLength",
          "termsAcceptanceRequired",
          "privacyAcceptanceRequired"
        ],
        "properties": {
          "publicRegistrationEnabled": {
            "type": "boolean"
          },
          "emailVerificationRequired": {
            "type": "boolean"
          },
          "passwordMinimumLength": {
            "type": "integer",
            "minimum": 12,
            "maximum": 128
          },
          "termsAcceptanceRequired": {
            "type": "boolean"
          },
          "privacyAcceptanceRequired": {
            "type": "boolean"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "password", "displayName"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "minLength": 1,
            "maxLength": 1024
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "acceptedTermsVersion": {
            "type": "string",
            "maxLength": 64
          },
          "acceptedPrivacyVersion": {
            "type": "string",
            "maxLength": 64
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "minLength": 1,
            "maxLength": 1024
          }
        }
      },
...
      "UserRole": {
        "type": "string",
        "enum": ["USER", "ADMIN"]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING_VERIFICATION",
          "ACTIVE",
          "DISABLED",
          "DELETION_PENDING",
          "DELETED"
        ]
      },
      "UserView": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "id",
          "email",
          "displayName",
          "role",
          "status",
          "emailVerified",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320
          },
          "displayName": {
            "type": "string",
            "maxLength": 80
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SessionView": {
        "type": "object",
        "additionalProperties": true,
        "required": ["user", "createdAt", "expiresAt"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AcceptedView": {
        "type": "object",
        "additionalProperties": true,
        "required": ["accepted"],
        "properties": {
          "accepted": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
    }
  }
}